bump -> cordova bump, split cordova task
