+++ config.xml
@@ -1,0 +1,1 @@
+cordova/config.xml
