+++ package.json
@@ -22,7 +22,7 @@
     "gulp-angular-filesort": "^1.0.4",
     "gulp-angular-templatecache": "~1.4.2",
     "gulp-autoprefixer": "~2.0.0",
-    "gulp-bump": "^0.1.13",
+    "gulp-cordova-bump": "^1.2.1",
     "gulp-csso": "~0.2.9",
     "gulp-filter": "~1.0.2",
     "gulp-flatten": "~0.0.4",
@@ -41,7 +41,6 @@
     "gulp-sourcemaps": "^1.3.0",
     "gulp-uglify": "~1.0.1",
     "gulp-useref": "~1.0.2",
-    "gulp-util": "^3.0.3",
     "http-proxy": "~1.7.0",
     "karma-coverage": "^0.2.7",
     "karma-jasmine": "~0.3.1",
@@ -52,8 +51,6 @@
     "optimist": "^0.6.1",
     "protractor": "~1.4.0",
     "require-dir": "~0.1.0",
-    "semver": "^4.2.2",
-    "through2": "^0.6.3",
     "uglify-save-license": "~0.4.1",
     "wiredep": "~2.2.0"
   },
